Moved argparser to separate __main__ file & simplified logic:
