+++ ssh_key_rotator/publisher/key_publisher.py
@@ -1,7 +1,6 @@
 from abc import ABC, abstractmethod
 import os
 import shutil
-from argparse import ArgumentParser
 import boto3
 from ssh_key_rotator.custom_keygen import (
     generate_private_public_key_in_file,
@@ -73,41 +72,3 @@ class LocalPublisher(Publisher):
             public_key_name=f"{self.user_id}-cert.pub",
         )
         return public_key.decode()
-
-
-def create_argument_parser() -> ArgumentParser:
-    argument_parser = ArgumentParser(
-        prog="key_publisher",
-        description="Creates public/private SSH keys and publishes the public key either locally or to S3 (default is S3)",
-        epilog="Thanks for using key_publisher! :)",
-    )
-
-    parser.add_argument("user")
-    parser.add_argument("hostname")
-    parser.add_argument("-ds", "--datastore", choices=["s3", "local"], default="s3")
-
-    options = parser.add_argument_group("Options")
-    options.add_argument(
-        "-ds (s3 | local)",
-        "--datastore (s3 | local)",
-        help="choose where to store the public key, on S3 or on the local system (default is S3)"
-    )
-
-    arguments = parser.add_argument_group("Required Arguments")
-    arguments.add_argument("user")
-    arguments.add_argument("hostname")
-
-    return argument_parser
-
-
-if __name__ == "__main__":
-    parser = create_argument_parser()
-    args = parser.parse_args()
-
-    if args.datastore == "local":  # If the user chose to store the public key locally
-        publisher = LocalPublisher(args.hostname, args.user)
-        publisher.publish_new_key()
-    else:  # If the user chose to store the public key on S3 or chose to default to S3
-        s3_bucket_name = input("Enter an S3 bucket name: ")
-        publisher = S3Publisher(s3_bucket_name, args.hostname, args.user)
-        publisher.publish_new_key()
